pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Andres Copeticona <andrescopeticona@outlook.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.7.9"
Flask = "2.0.3"
python-dotenv = "0.19.1"
flask-sqlalchemy = "2.5.1"
mysqlclient = "2.1.0"
jsonschema = "4.0.1"
pyjwt = "2.8.0"
pytz = "2023.3.post1"
boto3 = "1.28.51"
flask-cors = "4.0.0"
mediapipe = "0.8.5"
tensorflow = "2.4.1"
tensorflow-gpu = "2.4.1"
protobuf = "3.19.1"
pytest = "^7.4.3"
flask-socketio = "^5.3.6"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
sqlacodegen = { git = "https://github.com/agronholm/sqlacodegen.git" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
